adds working newspage, navbar, tourdates, start of media
The navbar has been fixed with some javascript and functions right now.
the newspage has working buttons and will get autoscroll soon.
the tourpage has been added and linked in the navigation
the mediapage has been added in basis and linked in the navigation
navigation is now soft-scrolling thanks to some jquery tricks.